Extracted component AssistantDropDown
